[refactor] refactor RuneSlotInfo
